refactor: change Fibonacci types from u32 to u64
Update all Fibonacci computations to use u64 instead of u32 for
better integer range support. This includes changes to the function
signatures, struct fields, and all call sites across the app, lib,
and prover modules. Also regenerate the RV64IM ELF binary with the
updated types.

+++ app/src/main.rs
@@ -6,10 +6,10 @@ use pico_sdk::io::{commit, read_as};
 
 pub fn main() {
     // Read inputs `n` from the environment
-    let n: u32 = read_as();
+    let n: u64 = read_as();
 
-    let a: u32 = 0;
-    let b: u32 = 1;
+    let a: u64 = 0;
+    let b: u64 = 1;
 
     // Compute Fibonacci values starting from `a` and `b`
     let (a_result, b_result) = fibonacci(a, b, n);
